Add the styling from amsterdam.github.org
Fixes #4

I've copied the design and general set up of Jekyll from the
amsterdam.github.io repository at commit b1a2827 and removed
all specific bits to make it possible to use the theme for other
and more generic projects.

https://github.com/Amsterdam/amsterdam.github.io/commit/b1a28272b1386c931e00c8705dc3b9d1daf1d748

Options for hiding the Logo or site title from the header are available.
This defaults to showing both.